backend/requirements.txt
django>=5.0,<6.0
djangorestframework>=3.15,<4.0
django-cors-headers>=4.3,<5.0
python-dotenv>=1.0,<2.0
anthropic>=0.39,<1.0
google-generativeai>=0.8,<1.0
Pillow>=12.1.1,<13.0
opencv-python-headless>=4.9,<5.0
pytesseract>=0.3.10,<1.0
pytest>=8.0,<9.0
pytest-django>=4.8,<5.0
